pyproject.toml
[project]
name = "matsemanns-streetview-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "dearpygui>=2.0.0",
    "pillow>=11.0.0",
    "pyright>=1.1.386",
    "pytest>=8.3.3",
    "spatialmedia",
    "tqdm>=4.66.5",
]

[tool.setuptools]
py-modules = ["matsemanns_streetview_tools"]

[tool.uv]

[tool.uv.sources]
spatialmedia = { git = "https://github.com/google/spatial-media", branch = "master" }
